third_party/WebKit/LayoutTests/editing/selection/node-removal-1.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div>^hello <span id="removeMe">removeMe|</span> world!</div>',
    selection => {
        selection.document.getElementById('removeMe').remove();
    },
    '<div>^hello | world!</div>'),
    'Verify selection after node removal');
</script>
